+++ libs/angular-modal-gallery/src/components/upper-buttons/upper-buttons-default.ts
@@ -5,15 +5,15 @@ import { Size } from '../../model/size.interface';
  * Default button size object
  * @type Size
  */
-export const defaultSize: Size = { height: 'auto', width: '30px' };
+export const KS_DEFAULT_SIZE: Size = { height: 'auto', width: '30px' };
 
 /**
  * Default close button object
  * @type ButtonConfig
  */
-export const closeButton: ButtonConfig = {
+export const KS_DEFAULT_BTN_CLOSE: ButtonConfig = {
   className: 'close-image',
-  size: defaultSize,
+  size: KS_DEFAULT_SIZE,
   type: ButtonType.CLOSE,
   title: 'Close this modal image gallery',
   ariaLabel: 'Close this modal image gallery'
@@ -22,9 +22,9 @@ export const closeButton: ButtonConfig = {
  * Default download button object
  * @type ButtonConfig
  */
-export const downloadButton: ButtonConfig = {
+export const KS_DEFAULT_BTN_DOWNLOAD: ButtonConfig = {
   className: 'download-image',
-  size: defaultSize,
+  size: KS_DEFAULT_SIZE,
   type: ButtonType.DOWNLOAD,
   title: 'Download the current image',
   ariaLabel: 'Download the current image'
@@ -33,9 +33,9 @@ export const downloadButton: ButtonConfig = {
  * Default exturl button object
  * @type ButtonConfig
  */
-export const extUrlButton: ButtonConfig = {
+export const KS_DEFAULT_BTN_EXTURL: ButtonConfig = {
   className: 'ext-url-image',
-  size: defaultSize,
+  size: KS_DEFAULT_SIZE,
   type: ButtonType.EXTURL,
   title: 'Navigate the current image',
   ariaLabel: 'Navigate the current image'
@@ -44,9 +44,9 @@ export const extUrlButton: ButtonConfig = {
  * Default refresh button object
  * @type ButtonConfig
  */
-export const refreshButton: ButtonConfig = {
+export const KS_DEFAULT_BTN_REFRESH: ButtonConfig = {
   className: 'refresh-image',
-  size: defaultSize,
+  size: KS_DEFAULT_SIZE,
   type: ButtonType.REFRESH,
   title: 'Refresh all images',
   ariaLabel: 'Refresh all images'
@@ -55,9 +55,9 @@ export const refreshButton: ButtonConfig = {
  * Default delete button object
  * @type ButtonConfig
  */
-export const deleteButton: ButtonConfig = {
+export const KS_DEFAULT_BTN_DELETE: ButtonConfig = {
   className: 'delete-image',
-  size: defaultSize,
+  size: KS_DEFAULT_SIZE,
   type: ButtonType.DELETE,
   title: 'Delete the current image',
   ariaLabel: 'Delete the current image'